[FE #47] RENAME : associated with ChatRoom
